added more constructors
