+++ .gitignore
@@ -6,3 +6,4 @@ target
 *.iml
 *~
 .idea
+.DS_Store
